cordova-sqlite-evcore-commoncore-free 0.0.1
NOTE that this version branch is currently known to require
slightly older Cordova CLI version such as cordova@8,
cordova@7, or cordova@6 in order to install properly.

This is due to use of an obsolete Cordova module in the
scripts/beforePluginInstall.js hook script.

This is not an issue with recent versions of
cordova-sqlite-storage itself or any of the
other up-to-date plugin version branches.