cordova-sqlite-evcore-commoncore-free 0.0.2
uses cordova-sqlite-evcore-free-dependencies@0.9.2,
with SQLite 3.30.1 update

NOTE that this version branch is currently known to require
older Cordova CLI version such as cordova@8,
cordova@7, or cordova@6 in order to install properly.

This is due to use of an obsolete Cordova module in the
scripts/beforePluginInstall.js hook script.

This outdated version branch is temporarily maintained
in order to provide some urgent patch updates
that can be merged into some other plugin versions.

This is not an issue with recent versions of
cordova-sqlite-storage itself or any of the
other up-to-date plugin version branches.